[Rename]: myDashboard 대소문자 수정
